+++ packages/hydrogen/src/components/CartLineQuantity/tests/CartLineQuantity.test.tsx
@@ -4,26 +4,28 @@ import {CartLineProvider} from '../../CartLineProvider';
 import {CartLineQuantity} from '../CartLineQuantity.client';
 import {CART_LINE} from '../../CartLineProvider/tests/fixtures';
 
-it('displays the quantity', () => {
-  const wrapper = mount(
-    <CartLineProvider line={CART_LINE}>
-      <CartLineQuantity />
-    </CartLineProvider>
-  );
+describe('<CartLineQuantity />', () => {
+  it.skip('displays the quantity', () => {
+    const wrapper = mount(
+      <CartLineProvider line={CART_LINE}>
+        <CartLineQuantity />
+      </CartLineProvider>
+    );
 
-  expect(wrapper).toContainReactComponent('span', {
-    children: CART_LINE.quantity,
+    expect(wrapper).toContainReactComponent('span', {
+      children: CART_LINE.quantity,
+    });
   });
-});
 
-it('allows a custom tag', () => {
-  const wrapper = mount(
-    <CartLineProvider line={CART_LINE}>
-      <CartLineQuantity as="p" />
-    </CartLineProvider>
-  );
+  it.skip('allows a custom tag', () => {
+    const wrapper = mount(
+      <CartLineProvider line={CART_LINE}>
+        <CartLineQuantity as="p" />
+      </CartLineProvider>
+    );
 
-  expect(wrapper).toContainReactComponent('p', {
-    children: CART_LINE.quantity,
+    expect(wrapper).toContainReactComponent('p', {
+      children: CART_LINE.quantity,
+    });
   });
 });