+++ packages/hydrogen/src/components/CartProvider/CartProvider.client.tsx
@@ -65,6 +65,9 @@ import {CART_ID_STORAGE_KEY} from './constants';
 import {CartFragmentFragment} from './graphql/CartFragment';
 import {CartQueryQuery, CartQueryQueryVariables} from './graphql/CartQuery';
 
+import {useServerState} from '../../foundation/useServerState';
+import {ServerStateContextValue} from '../../foundation';
+
 function cartReducer(state: State, action: CartAction): State {
   switch (action.type) {
     case 'cartFetch': {
@@ -252,6 +255,9 @@ export function CartProvider({
    */
   cart?: CartFragmentFragment;
 }) {
+  const {serverState} = useServerState() as ServerStateContextValue;
+  const countryCode = serverState?.country?.isoCode;
+
   const initialStatus: State = cart
     ? {status: 'idle', cart: cartFromGraphQL(cart)}
     : {status: 'uninitialized'};
@@ -259,15 +265,19 @@ export function CartProvider({
     (state, dispatch) => cartReducer(state, dispatch),
     initialStatus
   );
-  const fetch = useCartFetch();
+  const fetchCart = useCartFetch();
 
   const cartFetch = useCallback(
     async (cartId: string) => {
       dispatch({type: 'cartFetch'});
 
-      const {data} = await fetch<CartQueryQueryVariables, CartQueryQuery>({
+      const {data} = await fetchCart<CartQueryQueryVariables, CartQueryQuery>({
         query: CartQuery,
-        variables: {id: cartId, numCartLines},
+        variables: {
+          id: cartId,
+          numCartLines,
+          country: countryCode,
+        },
       });
 
       if (!data?.cart) {
@@ -278,7 +288,7 @@ export function CartProvider({
 
       dispatch({type: 'resolve', cart: cartFromGraphQL(data.cart)});
     },
-    [fetch, numCartLines]
+    [fetchCart, numCartLines, countryCode]
   );
 
   const cartCreate = useCallback(
@@ -287,7 +297,14 @@ export function CartProvider({
 
       onCreate?.();
 
-      const {data, error} = await fetch<
+      if (countryCode && !cart.buyerIdentity?.countryCode) {
+        if (cart.buyerIdentity == null) {
+          cart.buyerIdentity = {};
+        }
+        cart.buyerIdentity.countryCode = countryCode;
+      }
+
+      const {data, error} = await fetchCart<
         CartCreateMutationVariables,
         CartCreateMutation
       >({
@@ -295,6 +312,7 @@ export function CartProvider({
         variables: {
           input: cart,
           numCartLines,
+          country: countryCode,
         },
       });
 
@@ -317,7 +335,7 @@ export function CartProvider({
         );
       }
     },
-    [onCreate, fetch, numCartLines]
+    [onCreate, fetchCart, numCartLines, countryCode]
   );
 
   const addLineItem = useCallback(
@@ -325,7 +343,7 @@ export function CartProvider({
       if (state.status === 'idle') {
         dispatch({type: 'addLineItem'});
         onLineAdd?.();
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartLineAddMutationVariables,
           CartLineAddMutation
         >({
@@ -334,6 +352,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             lines: lines,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -352,7 +371,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, numCartLines, onLineAdd]
+    [fetchCart, numCartLines, onLineAdd, countryCode]
   );
 
   const removeLineItem = useCallback(
@@ -362,7 +381,7 @@ export function CartProvider({
 
         onLineRemove?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartLineRemoveMutationVariables,
           CartLineRemoveMutation
         >({
@@ -371,6 +390,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             lines: lines,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -389,7 +409,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onLineRemove, numCartLines]
+    [fetchCart, onLineRemove, numCartLines, countryCode]
   );
 
   const updateLineItem = useCallback(
@@ -399,7 +419,7 @@ export function CartProvider({
 
         onLineUpdate?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartLineUpdateMutationVariables,
           CartLineUpdateMutation
         >({
@@ -408,6 +428,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             lines: lines,
             numCartLines,
+            country: countryCode,
           },
         });
         if (error) {
@@ -425,7 +446,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onLineUpdate, numCartLines]
+    [fetchCart, onLineUpdate, numCartLines, countryCode]
   );
 
   const noteUpdate = useCallback(
@@ -435,7 +456,7 @@ export function CartProvider({
 
         onNoteUpdate?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartNoteUpdateMutationVariables,
           CartNoteUpdateMutation
         >({
@@ -444,6 +465,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             note: note,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -462,7 +484,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onNoteUpdate, numCartLines]
+    [fetchCart, onNoteUpdate, numCartLines, countryCode]
   );
 
   const buyerIdentityUpdate = useCallback(
@@ -472,15 +494,16 @@ export function CartProvider({
 
         onBuyerIdentityUpdate?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartBuyerIdentityUpdateMutationVariables,
           CartBuyerIdentityUpdateMutation
         >({
           query: CartBuyerIdentityUpdate,
           variables: {
             cartId: state.cart.id!,
-            buyerIdentity: buyerIdentity,
+            buyerIdentity,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -499,7 +522,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onBuyerIdentityUpdate, numCartLines]
+    [fetchCart, onBuyerIdentityUpdate, numCartLines, countryCode]
   );
 
   const cartAttributesUpdate = useCallback(
@@ -509,7 +532,7 @@ export function CartProvider({
 
         onAttributesUpdate?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartAttributesUpdateMutationVariables,
           CartAttributesUpdateMutation
         >({
@@ -518,6 +541,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             attributes: attributes,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -536,7 +560,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onAttributesUpdate, numCartLines]
+    [fetchCart, onAttributesUpdate, numCartLines, countryCode]
   );
 
   const discountCodesUpdate = useCallback(
@@ -549,7 +573,7 @@ export function CartProvider({
 
         onDiscountCodesUpdate?.();
 
-        const {data, error} = await fetch<
+        const {data, error} = await fetchCart<
           CartDiscountCodesUpdateMutationVariables,
           CartDiscountCodesUpdateMutation
         >({
@@ -558,6 +582,7 @@ export function CartProvider({
             cartId: state.cart.id!,
             discountCodes: discountCodes,
             numCartLines,
+            country: countryCode,
           },
         });
 
@@ -576,7 +601,7 @@ export function CartProvider({
         }
       }
     },
-    [fetch, onDiscountCodesUpdate, numCartLines]
+    [fetchCart, onDiscountCodesUpdate, numCartLines, countryCode]
   );
 
   const didFetchCart = useRef(false);
@@ -591,6 +616,14 @@ export function CartProvider({
       cartFetch(localStorage.getItem(CART_ID_STORAGE_KEY)!);
     }
   }, [cartFetch, state]);
+
+  useEffect(() => {
+    if (state.status !== 'idle') {
+      return;
+    }
+    buyerIdentityUpdate({countryCode}, state);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [countryCode]);
 
   const cartContextValue = useMemo(() => {
     return {