+++ packages/hydrogen/src/components/CartProvider/graphql/CartAttributesUpdateMutation.ts
@@ -5,6 +5,7 @@ export type CartAttributesUpdateMutationVariables = Types.Exact<{
   attributes: Array<Types.AttributeInput> | Types.AttributeInput;
   cartId: Types.Scalars['ID'];
   numCartLines?: Types.Maybe<Types.Scalars['Int']>;
+  country?: Types.Maybe<Types.CountryCode>;
 }>;
 
 export type CartAttributesUpdateMutation = {__typename?: 'Mutation'} & {