+++ packages/hydrogen/src/components/CartProvider/graphql/CartCreateMutation.ts
@@ -4,6 +4,7 @@ import {CartFragmentFragment} from './CartFragment';
 export type CartCreateMutationVariables = Types.Exact<{
   input: Types.CartInput;
   numCartLines?: Types.Maybe<Types.Scalars['Int']>;
+  country?: Types.Maybe<Types.CountryCode>;
 }>;
 
 export type CartCreateMutation = {__typename?: 'Mutation'} & {