+++ packages/hydrogen/CHANGELOG.md
@@ -7,6 +7,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- feat: use country server state in cart for the inContext directive
 - fix: update interaction prompt and interaction promp style attributes for Model3d
 - feat: use Image url field instead of deprecated originalSrc field
 - feat: switch to unstable API